Potential resolution to #18.
If this is actually what sets the order of data packs, then OCB will always be last which is not actually what I would want. Would want it last of mods but before player added ones. See no way of actually putting this where I want without looking though mod depdencies which would be very brittle. Old style of just creating the data pack and saying it should be enabled by default use to work but haven't tested it on any recent forge to see if the breakage was intentional or not.